// RodoApp-v2/types/PreNota.ts
// Definindo o formato dos dados recebidos
export type PreNota = {
  trim(Tipo: any): string
  Filial: string
  NF: string
  Status: string
  Fornecedor: string
  Emissao: string
  Inclusao: string
  Vencimento: string
  Valor: string
  Tipo: string
  Prioridade: string
  Usuario: string
  Obs: string
  ObsRev: string
  Revisao: string
  Rec: string
}